# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: TermiPay Admin Console API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all registered cardholders
      parameters:
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createUser
      tags: [users]
      summary: Register a new RFID card
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserBody"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get a user by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: User not found
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserBody"
      responses:
        "200":
          description: User updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: User not found
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: User deleted
        "404":
          description: User not found
...
components:
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: integer
        cardUid:
          type: string
        fullName:
          type: string
        contactNumber:
          type: string
        balance:
          type: number
        gcashLoadedTotal:
          type: number
        status:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - cardUid
        - fullName
        - contactNumber
        - balance
        - gcashLoadedTotal
        - status
        - createdAt
    CreateUserBody:
      type: object
      properties:
        cardUid:
          type: string
        fullName:
          type: string
        contactNumber:
          type: string
        initialBalance:
          type: number
      required:
        - cardUid
        - fullName
        - contactNumber
        - initialBalance
    UpdateUserBody:
      type: object
      properties:
        fullName:
          type: string
        contactNumber:
          type: string
        balance:
          type: number
        status:
          type: string
...