# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TermiPay Admin Console API
 * OpenAPI spec version: 0.1.0
 */

export interface User {
  id: number;
  cardUid: string;
  fullName: string;
  contactNumber: string;
  balance: number;
  gcashLoadedTotal: number;
  status: string;
  createdAt: Date;
}